Updated to   xkeyboard-config-1.7   pixman-0.16.2   libX11-1.3   libXinerama-1.1   xkbcomp-1.1.1   recordproto-1.14   xineramaproto-1.2   inputproto-2.0   compositeproto-0.4.1   xorg-server-1.7.0

+++ xorg-server/hw/xquartz/GL/visualConfigs.c
@@ -277,7 +277,7 @@ __GLXconfig *__glXAquaCreateVisualConfigs(int *numConfigsPtr, int screenNumber) 
     (c-1)->next = NULL;
 
     if (c - visualConfigs != numConfigs) {
-        FatalError("numConfigs calculation error in setVisualConfigs!  numConfigs is %d  i is %d\n", numConfigs, c - visualConfigs);
+        FatalError("numConfigs calculation error in setVisualConfigs!  numConfigs is %d  i is %d\n", numConfigs, (int)(c - visualConfigs));
     }
     
     freeGlCapabilities(&caps);
